+++ ibllib/atlas/atlas.py
@@ -383,7 +383,7 @@ class BrainAtlas:
         return extent
 
     def slice(self, coordinate, axis, volume='image', mode='raise', region_values=None,
-              mapping="Allen"):
+              mapping="Allen", bc=None):
         """
         :param coordinate: float
         :param axis: xyz convention:  0 for ml, 1 for ap, 2 for dv
@@ -428,6 +428,10 @@ class BrainAtlas:
             return _take(self.image, index, axis=self.xyz2dims[axis])
         elif volume in ['surface', 'edges']:
             return _take(self.surface, index, axis=self.xyz2dims[axis])
+        elif volume == 'volume':
+            if bc is not None:
+                index = bc.xyz2i(np.array([coordinate] * 3))[axis]
+            return _take(region_values, index, axis=self.xyz2dims[axis])
 
     def plot_cslice(self, ap_coordinate, volume='image', mapping='Allen', **kwargs):
         """
@@ -780,11 +784,12 @@ class AllenAtlas(BrainAtlas):
         Converts coordinates to the CCF coordinates, which is assumed to be the cube indices
         times the spacing.
         :param xyz: mlapdv coordinates in um, origin Bregma
-        :param ccf_order: 'mlapdv' (ibl) or 'apdvml' (Allen mcc vertices)
-        :return: coordinates in um (mlapdv by default), origin is the front left top corner
-         of the data volume
+        :param ccf_order: order that you want values returned 'mlapdv' (ibl) or 'apdvml'
+        (Allen mcc vertices)
+        :return: coordinates in CCF space um, origin is the front left top corner of the data
+        volume, order determined by ccf_order
         """
-        ordre = self._ccf_order(ccf_order, reverse=True)
+        ordre = self._ccf_order(ccf_order)
         ccf = self.bc.xyz2i(xyz, round=False) * float(self.res_um)
         return ccf[..., ordre]
 
@@ -792,11 +797,13 @@ class AllenAtlas(BrainAtlas):
         """
         Converts coordinates from the CCF coordinates, which is assumed to be the cube indices
         times the spacing.
-        :param mlapdv coordinates in um, origin is the front left top corner of the data volume
-        :param ccf_order: 'mlapdv' (ibl) or 'apdvml' (Allen mcc vertices)
+        :param ccf coordinates in CCF space in um, origin is the front left top corner of the data
+        volume
+        :param ccf_order: order of ccf coordinates given 'mlapdv' (ibl) or 'apdvml'
+        (Allen mcc vertices)
         :return: xyz: mlapdv coordinates in um, origin Bregma
         """
-        ordre = self._ccf_order(ccf_order)
+        ordre = self._ccf_order(ccf_order, reverse=True)
         return self.bc.i2xyz((ccf[..., ordre] / float(self.res_um)))
 
     @staticmethod