+++ ibllib/tests/test_ephys.py
@@ -57,4 +57,4 @@ class TestFpgaTask(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    unittest.main(exit=False)
+    unittest.main(exit=False, verbosity=2)