+++ oneibl/tests/test_one.py
@@ -49,51 +49,6 @@ class TestOneSetup(unittest.TestCase):
         self.assertTrue(self.pars_file.exists())
 
 
-class TestOneOffline(unittest.TestCase):
-
-    def setUp(self) -> None:
-        # init: create a temp directory and copy the fixtures
-        init_cache_file = Path(__file__).parent.joinpath('fixtures', '.one_cache.parquet')
-
-        # Create a temporary directory
-        self.test_dir = tempfile.TemporaryDirectory()
-
-        cache_dir = Path(self.test_dir.name)
-        shutil.copyfile(init_cache_file, cache_dir.joinpath(init_cache_file.name))
-
-        # test the constructor
-        self.one = ONE(offline=True)
-        self.assertTrue(self.one._cache.shape[1] == 14)
-
-        self.eid = 'cf264653-2deb-44cb-aa84-89b82507028a'
-
-    def test_one_offline(self) -> None:
-        # test the load with download false so it returns only file paths
-        one.list(self.eid)
-        dtypes = ['_spikeglx_sync.channels',
-                  '_spikeglx_sync.polarities',
-                  '_spikeglx_sync.times',
-                  '_iblrig_taskData.raw',
-                  '_iblrig_taskSettings.raw',
-                  'ephysData.raw.meta',
-                  'camera.times',
-                  'ephysData.raw.wiring']
-        one.load(self.eid, dataset_types=dtypes, dclass_output=False,
-                 download_only=True, offline=False)
-
-    def test_path_eid(self):
-        """Test `path_from_eid` and `eid_from_path` methods"""
-        eid = 'cf264653-2deb-44cb-aa84-89b82507028a'
-        # path from eid
-        session_path = self.one.path_from_eid(eid)
-        self.assertEqual(session_path.parts[-3:], ('clns0730', '2018-08-24', '002'))
-        # eid from path
-        self.assertEqual(eid, one.eid_from_path(session_path))
-
-    def tearDown(self) -> None:
-        self.test_dir.cleanup()
-
-
 class TestSearch(unittest.TestCase):
 
     def test_search_simple(self):
@@ -355,5 +310,50 @@ class TestMisc(unittest.TestCase):
         self.assertIsInstance(dsets, dict)
 
 
+class TestOneOffline(unittest.TestCase):
+
+    def setUp(self) -> None:
+        # init: create a temp directory and copy the fixtures
+        init_cache_file = Path(__file__).parent.joinpath('fixtures', '.one_cache.parquet')
+
+        # Create a temporary directory
+        self.test_dir = tempfile.TemporaryDirectory()
+
+        cache_dir = Path(self.test_dir.name)
+        shutil.copyfile(init_cache_file, cache_dir.joinpath(init_cache_file.name))
+
+        # test the constructor
+        self.one = ONE(offline=True)
+        self.assertTrue(self.one._cache.shape[1] == 14)
+
+        self.eid = 'cf264653-2deb-44cb-aa84-89b82507028a'
+
+    def test_one_offline(self) -> None:
+        # test the load with download false so it returns only file paths
+        one.list(self.eid)
+        dtypes = ['_spikeglx_sync.channels',
+                  '_spikeglx_sync.polarities',
+                  '_spikeglx_sync.times',
+                  '_iblrig_taskData.raw',
+                  '_iblrig_taskSettings.raw',
+                  'ephysData.raw.meta',
+                  'camera.times',
+                  'ephysData.raw.wiring']
+        one.load(self.eid, dataset_types=dtypes, dclass_output=False,
+                 download_only=True, offline=False)
+
+    def test_path_eid(self):
+        """Test `path_from_eid` and `eid_from_path` methods"""
+        eid = 'cf264653-2deb-44cb-aa84-89b82507028a'
+        # path from eid
+        session_path = self.one.path_from_eid(eid)
+        self.assertEqual(session_path.parts[-3:], ('clns0730', '2018-08-24', '002'))
+        # eid from path
+        self.assertEqual(eid, one.eid_from_path(session_path))
+
+    def tearDown(self) -> None:
+        self.test_dir.cleanup()
+
+
 if __name__ == '__main__':
-    unittest.main(exit=False)
+    unittest.main(exit=False, verbosity=2)