+++ release_notes.md
@@ -1,3 +1,10 @@
+## Release Notes 1.10
+### Release Notes 1.10.0 - 2021-05-04
+- TaskQC: exclude stim_freeze from overall session qc
+- Get modality from task type to include personal projects
+- Bugfix Atlas ccf coordinates order
+- Tool to label Critical sessions / insertions reasons
+
 ## Release Notes 1.9
 ### Release Notes 1.9.1 - 2021-04-19
 - Successful ONE setup on instantiation when params file doesn't exist